requirements.txt
pyyaml
goodreads
oauth2
